+++ packages/quantic/playwright/utils/requests.ts
@@ -22,9 +22,12 @@ export const analyticsEventsUrlRegex = /\/rest\/organizations\/.+?\/events\/v1/;
 export const searchQuickviewRequestRegex = /\/rest\/search\/v2\/html.*/;
 export const insightQuickviewRequestRegex =
   /\/rest\/organizations\/.*\/insight\/v1\/configs\/.*\/quickview/;
-
 export const rgaEvaluationsRequestRegex =
   /\/rest\/organizations\/.*\/answer\/v1\/configs\/.*\/evaluations/;
+export const rgaGenerateRequestRegex =
+  /\/rest\/organizations\/.*\/answer\/v1\/configs\/.*\/generate/;
+export const insightRgaGenerateRequestRegex =
+  /\/rest\/organizations\/.*\/insight\/v1\/configs\/.*\/generate/;
 
 /**
  * Indicates whether the specified request corresponds to a Search Usage Analytics request.
@@ -47,12 +50,13 @@ export function isUaCustomEvent(request: Request): boolean {
 }
 
 /**
- * Indicates whether the specified request corresponds to an RGA evaluation request.
+ * Indicates whether the specified request corresponds to a RGA Evaluation request.
  * @param request The request to check.
  */
 export function isRgaEvaluationRequest(request: Request): boolean {
   return (
-    request.method() === 'POST' && rgaEvaluationsRequestRegex.test(request.url())
+    request.method() === 'POST' &&
+    rgaEvaluationsRequestRegex.test(request.url())
   );
 }